Implement Stack status and action properties
This splits the stack_status back to an action and status,
as it is represented internally in heat-engine.

It assumes there will not be an action which contains an _.

This change is needed for code (such as tempest) which assumes a
convention of all openstack client resources having a status property.

Change-Id: If00db5508bf8469b1106cac20b8fc10f7e1a854d