add output-list and output-show
This patch adds the "output-list" and "output-show" commands to heat.
The first prints a table of available outputs to stdout, and the second
prints a single output_value to stdout.

All outputs are serialized to JSON.

This patch includes tests for the new commands as well as updated
documentation.

Implements blueprint: stack-outputs-in-shell

Change-Id: Iaade16d043b2c42d14f642fde17419eaa07d0ab8
Closes-Bug: 1258622