REST method to fetch deployments metadata for a server
This REST method returns a grouped collection of deployment resource
metadata which a server uses as its configuration inputs.

Previously the index method was overloaded to perform this function too.

Currently this will be called by heat-engine when the server
resource metadata is polled, but it should eventually be possible
for servers to call this method directly so that a stack does not
need to be parsed every time a server polls for metadata.

partial blueprint hot-software-config

Change-Id: Ie81e413459eaf7a24658be84152d88228f42c6b6