Output warnings for deprecated commands
There are various options which have been deprecated for a long
time but with no user-visble indication of the deprecated status.
So log a warning to encourage the user to switch to the new
interfaces instead of just having a strongly-worded docstring ;)

Hopefully this means we can, at some future point, justifiably
remove the duplicate deprecated functions.

Change-Id: Ia5d791de99e7238645beb7d7650d811225d9d9c6