Add rel field to links display
Currently the "rel" keys from the links list are not displayed, so
it's not that clear which link refers to what unless you're familiar
with heat.  So add a suffix which mentions the rel so e.g users can
see "nested" there and tell the link refers to a nested stack.

Change-Id: I5abe7f7b51c290bd9d8daaccb7f9a842af898f58