Add --nested-depth option to event-list
Adds logic to mimic the resource-list nested-depth option for events.

Note that this is pretty inefficient, and in future we should add an
API for nested_depth to the events API, but I want this to work for
kilo heat, so this interim implementation will work for kilo, then
I'll look at an alternative (faster) API-side implementation for
Liberty, which will maintain the same CLI interfaces.

Change-Id: I76c60ab5b79af9c477af07d5690b8ca6ca4da388