Add hook-poll function to check if a stack has pending hooks
Currently using the breakpoints/hooks functionality on large trees of
stacks is a bit inconvenient, because you may set hooks in nested stacks
so the events aren't easily visible via the top-level event interfaces.

This has been addressed via the new --nested-depth option to event-list,
but it's still hard to determine (programatically e.g from a script) if
there's pending hook which requires clearing.

So this patch introduces an initial implementation of a specialized
event-list function "hook-poll", which filters the events and displays
only those events which have pending hooks (e.g those which have yet
to be signalled to clear them).

I expect the efficiency of the implementation can be much improved in the
future if we add a propert nested_depth argument to the heat events API,
and/or add a hook API which enables easier introspection of hook status.
The CLI interface should be reasonable in the event such rework happens,
but the current (slow) implementation will work with the API interfaces
we have available now in kilo Heat.

Change-Id: I71b19202ab29f44e5c09b4ee04be4aeaea038c28