Fix sort column for event-list with nested-depth
The non-nested-depth case can rely on the sort ordering from the
API response, but because we mangle together the response from
several event-list API calls in the nested-depth call, it's necessary
to ensure the sort ordering is correct by sorting on event_time.

Currently, we're sorting on the wrong column, because of the logic
which inserts a new column heading after we look up the index,
but switching to the right column still relies on the prettytable
sortby_index, which doesn't work with the "start" slice option,
if you want post-sort slicing (such as will be required to make
--marker and --limit work with the --nested-depth option, in a
later patch).

So, instead, pre-sort the list before passing to print_list, which
solves the immediate problem of wrong sort order, and provides
an easier step towards enabling marker/limit.

Change-Id: Ifc6c8875cfdc58c0185fbffe7d00a45a598d67bd