Make hook type optional for hook-clear
We can derive the correct hook type (and do some sanity assertions,
as in hook-poll) via the stack status, so it's slightly easier,
particularly for scripts, if you don't need to explicitly specify
the hook type.

The old options are retained for compatibility, and I guess they
will make the call slightly more efficient.

Change-Id: I622f13e10d17e31ea2a6a2d7647c3e6ef3b085bf