OpenStackClient plugin for stack show
This change was derived from Id188cdc0e97b480875a7626443f68e69863f0647
which added stack support to python-openstackclient, but adapting it
to be an openstackclient plugin which lives in the python-heatclient
tree.

This change only implements the "openstack stack show". Subsequent
commands will remain in WIP changes at the end of this series until they
are ready to merge.

The stack show formatting has the following behaviour:
- specifying key order for important values, but adding all other
  stack keys in default order (future proofing for new values)
- exclude template_description, its a dupe of 'description'
- complex values like parameters, outputs and links get a special formatter
  depending on the output format ('table' formats them as yaml,
  'shell', 'value', 'html' formats them as json)

Co-Authored-By: Ryan S. Brown <rybrown@redhat.com>
Co-Authored-By: Rico Lin <rico.l@inwinstack.com>
Change-Id: I3096b94146a94d184c29b8c7c9f6c032eed5281d