Modify test templates and the path to subunit2html
This patch contains the following two fixes
because the gate job fails if they are not applied
at the same time.

1) Use OS::Heat::TestResource in test templates

Use OS::Heat::TestResource instead of OS::Nova::Server
in heat_minimal.yaml and heat_minimal_hot.yaml

2) Update path to subunit2html in post_test_hook

Per:

http://lists.openstack.org/pipermail/openstack-dev/2015-August/072982.html

The location of subunit2html changed on the images in the gate
so update the path used in the post_test_hook.

Long-term we should just use what's in devstack-gate.

Change-Id: I67acc3145719c3de1199e36c6667d975652d262f
Closes-Bug: #1492115
Closes-Bug: #1491646