Fix error when listing stacks
Previously the stack list command would error out if any stacks were
present. This was due to the action code attempting to index on a
stack object, rather than grabbing its attributes. This change switches
the action code to use the proper utility for a resource instead of
a dict.

Change-Id: Ia6b73d5c1ba7f7e3d7bc5493167da81062102b42
Blueprint: heat-support-python-openstackclient