Update --template-object to support nested stacks
This patch updates heatclient so that it supports creating
stacks from an object (Swift container) which contain nested
stack relative links to nested stack URLs.

The motivation here:

We would like to be able to use heatclient to create stacks directly
from Swift containers in TripleO. The --template-object support already
exists and is great... it just needs to support loading the associated
nested stacks as well.

Change-Id: I53ac83e9f9985224d19993935467c03611aaef6a
Closes-bug: #1532326