Implement "openstack stack failures list"
A command which prints failed resources in the stack to help debug
stack issues.

Failed resources are added by recursing from the top level stack into
failed nested stack resources. A failed nested stack resource is only
added to the failed list if it contains no failed resources so that the
user only sees root-cause failed resources.

As an added convenience to debugging deployment resources, deployment
outputs deploy_stdout and deploy_stderr are printed as well (the full
outputs are printed when --long is specified).

The output is a yamlish format where the key is a dot-delimited resource
name path and the value is a dict of resource information.

Change-Id: Ifa44175d8dbab2f7b62691b839e1696e3af7f4f3
Closes-Bug: #1585820