Clean up osc fakes and mocks
The original fakes included a lot of unneeded pasted code from
openstackclient. This change replaces the client with a single
MagicMock, and deletes unused fakes classes.

Since the base client is a magic mock, a significant amount of mock
object creation has been replaced with the magic equivalent.

This change actually revealed a faulty test,
test_deployment_config_delete_failed should be asserting that deleting
the derived config is best-effort and does not raise an error.

Change-Id: I11be97603b2d42ea70dde7d936e11d15e7f5709e