Register a log formatter for stack event list
Currently "stack event list" overrides the value formatter to use for
the default log formatting style. This change creates a formatter
namespace for event list so that it can have its own custom list of
formatters.

This means that the 'log' formatter is now registered and is the
default. The formatter expects event objects rather than dicts, and
so ignores the columns argument.

The intention is that this will be the first of more custom formatters
for event list, making the effort to establish this pattern worth it.

Change-Id: I22f884f871d7467c8abadec477b3e732eb8c2150