Fix obscure error message when no template given to OSC
This is an internal-only fix, as it does not change any external APIs. A
bit of obfuscation was occurring in the
'template_utils.process_template_path' function call. First we try to
retrieve the template from a urllib call. If that fails, then we try to
make a request with the custom request_object function. If the
request_object function failed to retrieve the template as well, it
would report back 'ERROR: None'. I left both of the calls in this
function, but I made the first exception bubble up back to the caller,
since it is more helpful to the user.

The new error message should report back the following when a template
cannot be found:

<urlopen error [Errno 2] No such file or directory:
'/root/python-heatclient/test.yaml'>

Closed Bug: #1570229

Change-Id: I29e19172322c18559aa476744f7ff3530e689bd0