Add parameter_merge_startegies section
This adds parameter_merge_strategies section to environment.
This also adds the missing section parameter_encrypted_names.
We use a server side module and a client side module that have
deviated over a period of time. This patch syncs them. We probably
should move these to a library and use use them for both server and
client in the future.

Change-Id: I9b20a7cedd416b6d8b788a30f66e472b85babf4a
Blueprint: environment-merging