Don't resolve outputs when polling events
In poll_for_events, we retrieve the stack to check its status. As it
stands we resolve the outputs as well, which is costly and useless in
this scenario. We can simpley get the stack to retrieve its status
without outputs.

Change-Id: Idc2327f6be15a679d502b0cd3e9d8980fdf1e22a
Related-Bug: #1638908
(cherry picked from commit cf2ecd453a82a0eab77e8645d27e74182cec942c)