fix tox python3 overrides
We want to default to running all tox environments under python 3, so
set the basepython value in each environment.

We do not want to specify a minor version number, because we do not
want to have to update the file every time we upgrade python.

We do not want to set the override once in testenv, because that
breaks the more specific versions used in default environments like
py35 and py36.

Change-Id: Ia20d2ef56af2da00b374b3a3e5d79d7228c6a763
Signed-off-by: Doug Hellmann <doug@doughellmann.com>