Require int value for --limit option
Currently the value passed to the --limit option is not validated as
part of argument parser and results in an ugly error.

$ openstack stack list --limit a
invalid literal for int() with base 10: 'a'
$

This enables the validation in argument parser to require integer,
which shows more clear error.

$ openstack stack list --limit a
openstack stack list: error: argument --limit: invalid int value: 'a'
$

Change-Id: I17397bc202a08de63b715fa15d0f33754393c226